Separate the layout
