+++ src/app/users/user-single/user-single.component.ts
@@ -5,6 +5,11 @@ import { Component, OnInit } from '@angular/core';
   templateUrl: './user-single.component.html',
   styleUrls: ['./user-single.component.scss']
 })
-export class UserSingleComponent {
+export class UserSingleComponent implements OnInit {
+  constructor () {
 
+  }
+  ngOnInit () {
+    
+  }
 }
